+++ sensor/common/sensor/sensor.go
@@ -19,6 +19,7 @@ import (
 	"github.com/stackrox/rox/pkg/grpc/routes"
 	"github.com/stackrox/rox/pkg/listeners"
 	"github.com/stackrox/rox/pkg/logging"
+	"github.com/stackrox/rox/pkg/mtls"
 	"github.com/stackrox/rox/pkg/mtls/verifier"
 	"github.com/stackrox/rox/pkg/orchestrators"
 	"github.com/stackrox/rox/pkg/retry"
@@ -128,7 +129,7 @@ func (s *Sensor) Start() {
 	// Start up connections.
 	log.Infof("Connecting to Central server %s", s.centralEndpoint)
 	var err error
-	s.centralConnection, err = clientconn.AuthenticatedGRPCConnection(s.centralEndpoint, clientconn.Central)
+	s.centralConnection, err = clientconn.AuthenticatedGRPCConnection(s.centralEndpoint, mtls.CentralSubject)
 	if err != nil {
 		log.Fatalf("Error connecting to central: %s", err)
 	}